+++ src/Http/View/InjectJavascriptVariables.php
@@ -43,6 +43,7 @@ class InjectJavascriptVariables
             'user' => $this->authentication->getUser(),
             'group' => $this->authentication->getGroup(),
             'role' => $this->authentication->getRole(),
+            'position' => $this->authentication->hasRole() ? $this->authentication->getRole()->position() : null,
             'activity' => ($this->request->route()->hasParameter('activity_slug') ? $this->request->route('activity_slug') : null),
             'activity_instance' => $this->getActivityInstance(),
             'module_instance' => ($this->request->route()->hasParameter('module_instance_slug') ? $this->request->route('module_instance_slug') : null)