+++ tests/Settings/Definition/GlobalSettingTest.php
@@ -124,7 +124,7 @@ class GlobalSettingTestDummyGlobalSetting extends GlobalSetting
      */
     public function fieldOptions(): Field
     {
-        return \FormSchema\Generator\Field::input($this->inputName())->inputType('text')->getSchema();
+        return \FormSchema\Generator\Field::textInput($this->inputName());
     }
 
     /**