+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Resolve Transformer contract rather than using VFGTransformer.
+- PR-344: Use new form schema
+- Fix client caching not properly caching response
+
+### Added
+
+- Add position to js variables
+- `participant-activity` and `participant-module` middleware groups
+- Migrations to update settings
+- Added a title prefix to resource
+- Name the main module route 'module.admin' and 'module.participant'
+
+## [5.0.0] - (11/08/2021)
+
 ### Removed
 - Cookies warning blade template is no longer included with the SDK